+++ explorer/src/components/common/NFTArt.tsx
@@ -242,6 +242,7 @@ export const ArtContent = ({
   uri,
   animationURL,
   files,
+  data
 }: {
   metadata: MetadataData;
   category?: MetaDataJsonCategory;
@@ -250,10 +251,9 @@ export const ArtContent = ({
   uri?: string;
   animationURL?: string;
   files?: (MetadataJsonFile | string)[];
+  data: MetadataJson | undefined
 }) => {
   const id = pubkeyToString(pubkey);
-
-  const { data } = useExtendedArt(id, metadata);
 
   if (pubkey && data) {
     uri = data.image;
@@ -356,54 +356,4 @@ export const useCachedImage = (uri: string) => {
   }, [uri, setCachedBlob, fetchStatus, setFetchStatus]);
 
   return { cachedBlob };
-};
-
-export const useExtendedArt = (id: string, metadata: MetadataData) => {
-  const [data, setData] = useState<MetadataJson>();
-
-  useEffect(() => {
-    if (id && !data) {
-      if (metadata.data.uri) {
-        const uri = metadata.data.uri;
-
-        const processJson = (extended: any) => {
-          if (!extended || extended?.properties?.files?.length === 0) {
-            return;
-          }
-
-          if (extended?.image) {
-            extended.image = extended.image.startsWith("http")
-              ? extended.image
-              : `${metadata.data.uri}/${extended.image}`;
-          }
-
-          return extended;
-        };
-
-        try {
-          fetch(uri)
-            .then(async (_) => {
-              try {
-                const data = await _.json();
-                try {
-                  localStorage.setItem(uri, JSON.stringify(data));
-                } catch {
-                  // ignore
-                }
-                setData(processJson(data));
-              } catch {
-                return undefined;
-              }
-            })
-            .catch(() => {
-              return undefined;
-            });
-        } catch (ex) {
-          console.error(ex);
-        }
-      }
-    }
-  }, [id, data, setData, metadata.data.uri]);
-
-  return { data };
 };
