+++ explorer/src/scss/_solana.scss
@@ -406,6 +406,11 @@ p.updated-time {
   width: 150px;
 }
 
+.external-url-button {
+  margin-left:0.4rem !important;
+  border-radius:0.25rem !important;
+}
+
 .creator-dropdown-entry {
   flex-flow: row wrap;
 }
